Introduce MacroExpansionAttachment#expandee, dedup MacroExpansionOf
